services/ticket/requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
pymysql
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
httpx
